feat(lib): generator
Create a `PDUGenerator` in its own file, and also its tests in a dedicated file. It shall have a bunch of methods to configure some common fields and encodings. It also has a `generate` method that accept a data from which to take bytes and an optional content-type. It shall yield chunks and/or message.