feat(deframer): heuristic
In case it is not possible to peek into the header, add an heuristic that uses all latest unfinished announcement post-boundary encodings to try to decode the just received bytes.

Assume all encodings (compressions and crcs) could be used anywhere. Gzip is a good example of encodings disabling one to peek headers.

+ plan review