feat: Reed-Solomon
Add support of Reed-Solomon encoding thanks to `reedsolo` library. It is parametize with chunk-size, egg 255, and actual data-size, eg 233. The encoding label of this example would be  `rs(255,233)` meaning data is splitted in chunks of 233 bytes, and after Reed-Solomon it would make 255 bytes of data. Update tests, Deframer, and PDUGenerator accordingly.