feat: send udp CLI
Add the implementation of a CLI that take as argument a filepath, an ip address, a UDP port, a src-callsign, an encodings list, and an optional annoucement_encodings list. It will leverage the PDUGenerator to send as many UDP frames as HQFBP PDU generated. Try to guess the file mimetype.

Add a test of the CLI for encodings like `rs(255,233)`.